requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
reportlab
python-pptx
requests
beautifulsoup4
